+++ src/entities.rs
@@ -1,4 +1,5 @@
 mod full;
+mod internal;
 mod threshold;
 
 pub use full::{AuxInfo, KeyShare, KeyShareChange};
@@ -7,3 +8,4 @@ pub use threshold::ThresholdKeyShare;
 pub(crate) use full::{
     AuxInfoPrecomputed, PublicAuxInfo, PublicAuxInfoPrecomputed, PublicAuxInfos, PublicKeyShares, SecretAuxInfo,
 };
+pub(crate) use internal::Sid;